style: move home page component
